# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Acesse SUS API Documantation",
    "description": "In this documantation you will be able to consult the API end-points and also test all availblw routes. Do not dorfet to register and carry out the authorization.",
    "contact": {
      "email": "joey_qpb@hotmail.com"
    },
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/API/1.0/MED/login": {
      "post": {
        "summary": "Login MED",
        "tags": ["/Doctor"],
        "description": "this route is responsible for loging in a user of type MED",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "email": { "type": "string" },
                "password": { "type": "string" }
              },
              "examples": {
                "MED Login": {
                  "value": {
                    "email": "medname@exemple.com",
                    "password": "thispassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MEDLogin"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    },
    "/API/1.0/MED/consulta/{id}": {
      "post": {
        "summary": "create medical consultation",
        "tags": ["/Doctor"],
        "security": [{ "bearerAuth": [] }],
        "description": "this route allows the doctor to create a new query in the database or update the previous query causing the previous query to be added to the 'history' field of the 'queries' collection",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "consulta id",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Consulta"
              },
              "examples": {
                "consulta": {
                  "type": "object",
                  "properties": {
                    "info": "information from the patient about the consultation",
                    "diagnostico": "medical diagnostic",
                    "prognostico": "medical prognosis",
                    "receita": "prescription",
                    "obs": "some medical observation"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "medical consultation created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Consulta"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "User Unauthorized"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    },
    "/API/1.0/MED/perfil": {
      "get": {
        "summary": "Doctor profile",
        "tags": ["/Doctor"],
        "security": [{ "bearerAuth": [] }],
        "description": "this route allows the doctors to have access to their information registered in the database",
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/MEDperfil"
                }
              }
            }
          },
          "401": {
            "description": "User Unauthorized"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    },
    "/API/1.0/MED/info_paciente_c/{cpf}": {
      "get": {
        "summary": "get patient profile",
        "tags": ["/Doctor"],
        "security": [{ "bearerAuth": [] }],
        "description": "this route gives doctors access to a specific patient registered in the database by cpf of patient",
        "parameters": [
          {
            "in": "path",
            "name": "cpf",
            "type": "string",
            "description": "patient cpf",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/PACperfil"
                }
              }
            }
          },
          "401": {
            "description": "User Unauthorized"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PACperfil": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "required": true
          },
          "email": {
            "type": "string",
            "required": true,
            "uniqueItems": true
          },
          "cpf": {
            "type": "string",
            "required": true,
            "uniqueItems": true
          },
          "rg": {
            "type": "string",
            "required": true,
            "uniqueItems": true
          },
          "posto": {
            "type": "string"
          },
          "foto": {
            "type": "file",
            "default": "https://res.cloudinary.com/dqeehrnvz/image/upload/v1670727739/Acesse%20Sus%20Pasta/default_img_ml07es.png"
          },
          "nomesocial": {
            "type": "string",
            "default": "Não Tem ou Não Informado"
          },
          "suscard": {
            "type": "string",
            "required": true
          },
          "nacionalidade": {
            "type": "string",
            "required": true
          },
          "dataNascimento": {
            "type": "string",
            "required": true
          },
          "cor": {
            "type": "string",
            "required": true
          },
          "sexo": {
            "type": "string",
            "required": true
          },
          "nomeDaMae": {
            "type": "string",
            "required": true
          },
          "tel": {
            "type": "string",
            "required": true
          },
          "pulmaoDoenca": {
            "type": "string",
            "required": true
          },
          "fumante": {
            "type": "string",
            "required": true
          },
          "usaAlcool": {
            "type": "string",
            "required": true
          },
          "usaDrogas": {
            "type": "string",
            "required": true
          },
          "hipertenso": {
            "type": "string",
            "required": true
          },
          "diabetes": {
            "type": "string",
            "required": true
          },
          "avcderrame": {
            "type": "string",
            "required": true
          },
          "infarto": {
            "type": "string",
            "required": true
          },
          "cardioDoenca": {
            "type": "string",
            "required": true
          },
          "problemaRins": {
            "type": "string",
            "required": true
          },
          "createdBy": {
            "type": "string"
          },
          "updateBy": {
            "type": "string"
          },
          "consulta": {
            "type": "Types.ObjectId"
          }
        }
      },
...
      "MEDLogin": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "name": { "type": "string" },
              "email": { "type": "string" },
              "_id": { "type": "string" },
              "role": { "type": "string" }
            }
          },
          "token": { "type": "string" }
        }
      },
      "MEDperfil": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "required": true
          },
          "email": {
            "type": "string",
            "required": true,
            "uniqueItems": true
          },
          "role": { "type": "string" },
          "cpf": {
            "type": "string",
            "required": true,
            "uniqueItems": true
          },
          "rg": {
            "type": "string",
            "required": true,
            "uniqueItems": true
          },
          "posto": {
            "type": "string"
          },
          "foto": {
            "type": "string",
            "default": "https://res.cloudinary.com/dqeehrnvz/image/upload/v1670727739/Acesse%20Sus%20Pasta/default_img_ml07es.png"
          },
          "CRM": {
            "type": "string",
            "required": true,
            "uniqueItems": true
          },
          "UF": {
            "type": "string",
            "enum": [
              "AC",
              "AL",
              "AP",
              "AM",
              "BA",
              "CE",
              "DF",
              "ES",
              "GO",
              "MA",
              "MT",
              "MS",
              "MG",
              "PA",
              "PB",
              "PR",
              "PE",
              "PI",
              "RJ",
              "RN",
              "RS",
              "RO",
              "RR",
              "SC",
              "SP",
              "SE",
              "TO",
              ".."
            ],
            "default": ".."
          },
          "especialidade": {
            "type": "string",
            "required": true
          },
          "consultas": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },

          "createdBy": {
            "type": "string"
          },
          "updateBy": {
            "type": "string"
          },
          "dataNascimento": {
            "type": "string",
            "required": true
          }
        }
      },
      "Consulta": {
        "type": "object",
        "properties": {
          "info": {
            "type": "string",
            "required": true
          },
          "diagnostico": {
            "type": "string",
            "required": true,
            "default": "Indefinido"
          },
          "prognostico": {
            "type": "string",
            "required": true,
            "default": "Indefinido"
          },
          "receita": {
            "type": "string",
            "required": true
          },
          "obs": {
            "type": "string",
            "required": true,
            "default": "Sem Observações a serem feitas"
          },
          "pacienteId": {
            "type": "string"
          },
          "pacienteNome": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date",
            "default": "new Date(Date.now())"
          },
          "createdBy": {
            "type": "Types.ObjectId",
            "ref": "Medico"
          },
          "medico": {
            "type": "string"
          },
          "medicoId": {
            "type": "Types.ObjectId",
            "ref": "Medico"
          },
          "updatedAt": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date"
            }
          },
          "historico": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "info": {
                  "type": "string",
                  "required": true
                },
                "diagnostico": {
                  "type": "string",
                  "required": true,
                  "default": "Indefinido"
                },
                "prognostico": {
                  "type": "string",
                  "required": true,
                  "default": "Indefinido"
                },
                "receita": {
                  "type": "string",
                  "required": true
                },
                "obs": {
                  "type": "string",
                  "required": true,
                  "default": "Sem Observações a serem feitas"
                },
                "pacienteId": {
                  "type": "string"
                },
                "pacienteNome": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date",
                  "default": "new Date(Date.now())"
                },
                "createdBy": {
                  "type": "Types.ObjectId",
                  "ref": "Medico"
                },
                "medico": {
                  "type": "string"
                },
                "medicoId": {
                  "type": "Types.ObjectId",
                  "ref": "Medico"
                },
                "updatedAt": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}